requirements.txt
certifi==2019.6.16
chardet==3.0.4
idna==2.8
numpy==1.16.4
picamera==1.13
Pillow==6.0.0
pkg-resources==0.0.0
pyzbar==0.1.8
requests==2.22.0
urllib3==1.25.3
